Remove `urw-core35-fonts' from dependencies list
Even though it was listed as an "optional" dependency, NPM downloaded it
indiscriminately each time it installed the `roff` module. The resulting
download is non-trivial in size, and it was originally believed that the
`optionalDependencies` were only installed if explicitly indicated by an
option on command-line.

Obviously, that wasn't the case...